# docs/swagger.yaml
definitions:
  middleware.ErrorResponse:
    description: Standard API error response
    properties:
      error:
        example: Invalid request
        type: string
      message:
        example: Detailed error message
        type: string
    type: object
  user.StandardResponse:
    description: Standard API response
    properties:
      message:
        example: Operation successful
        type: string
    type: object
  user.Tokens:
    description: Authentication tokens for user sessions.
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  user.UpdateRequest:
    properties:
      password:
        type: string
    required:
    - password
    type: object
  user.User:
    description: User model containing user information
    properties:
      email:
        example: user@example.com
        type: string
      name:
        example: John
        type: string
      password:
        example: SuperStrongPassword123
        type: string
      surname:
        example: Doe
        type: string
      username:
        example: johndoe
        type: string
    required:
    - email
    - name
    - password
    - surname
    - username
    type: object
  user.UserLogin:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  user.UserResponse:
    properties:
      created_at:
        example: "2020-01-01 12:00:00"
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: John
        type: string
      surname:
        example: Doe
        type: string
      updated_at:
        example: "2020-01-01 12:00:00"
        type: string
      username:
        example: johndoe
        type: string
    type: object
info:
  contact: {}
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password.
      parameters:
      - description: User credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: Access token is returned in response. Refresh token is set
            in a cookie.
          schema:
            $ref: '#/definitions/user.Tokens'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: User Login
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Generates a new access token using the provided refresh token.
      produces:
      - application/json
      responses:
        "200":
          description: New access token is returned in response.
          schema:
            $ref: '#/definitions/user.Tokens'
        "401":
          description: Refresh token required
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Refresh Access Token
      tags:
      - auth
  /auth/sign_up:
    post:
      consumes:
      - application/json
      description: Register a new user with email and password.
      parameters:
      - description: User registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.User'
      produces:
      - application/json
      responses:
        "200":
          description: 'status: ok'
          schema:
            $ref: '#/definitions/user.StandardResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: User Registration
      tags:
      - auth
  /health:
    get:
      description: Checks if the API is running.
      produces:
      - application/json
      responses:
        "200":
          description: 'status: healthy'
          schema:
            $ref: '#/definitions/user.StandardResponse'
      summary: Health Check
      tags:
      - system
  /user:
    delete:
      description: Deletes the currently logged-in user. Requires an access token
        in the `Authorization` header.
      responses:
        "200":
          description: 'message: User deleted successfully'
          schema:
            $ref: '#/definitions/user.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete User
      tags:
      - user
  /user/{id}:
    get:
      consumes:
      - application/json
      description: Fetches user details using their ID. Requires an access token in
        the `Authorization` header.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - Bearer: []
      summary: Get User
      tags:
      - user
  /user/logout:
    post:
      description: Logs out the authenticated user and invalidates their refresh token.
        Requires an access token in the `Authorization` header.
      responses:
        "200":
          description: 'message: Logged out successfully'
          schema:
            $ref: '#/definitions/user.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - Bearer: []
      summary: User Logout
      tags:
      - auth
  /user/update_password:
    put:
      consumes:
      - application/json
      description: Allows an authenticated user to update their password. Requires
        an access token in the `Authorization` header.
      parameters:
      - description: New password details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Password updated successfully'
          schema:
            $ref: '#/definitions/user.StandardResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      security:
      - Bearer: []
      summary: Update Password
      tags:
      - user
swagger: "2.0"